Remove compe support.
It's long deprecated/abandoned in favor of nvim-cmp.

We can add some extra nvim-cmp support at some point (including
maybe re-adding an abbreviation source) but for now remove this
so it's not confusing.

Closes: #140
